[TASK] Adjust the apache php-module based image.
This change provides the changes for the apache2 based
image. Beside the goal to integrate the same extension
as the original cli based images, some apache config
enhancement are added along with a configurable docroot
like the configurable apache run user and group settings.

This is needed, as we want to mount volumes with the exact
same path in the container like the path is on the host,
instead of the casual translation to `/var/www/html`.